Patch axios to allow custom configuration within interceptors
We have an issue in our services with axios and axios-retry. With the release of axios@0.19.0,
axios introduced a breaking change (without proper notification or mayor version release)
where custom properties are ignored when creating a request, and not passed to interceptors.
Custom properties are used by plugins like axios-retry, to keep the count of retries per request.
Since this change in June, our retry mechanism is completely broken and NEVER STOPS RETRYING.